backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.6
pydantic-settings==2.8.1
redis==5.2.1
qdrant-client==1.13.2
httpx==0.28.1
python-jose[cryptography]==3.5.0
python-multipart==0.0.22
orjson==3.11.7
beautifulsoup4==4.13.3
pytest==8.3.5
